backend/requirements.txt
asgiref==3.8.1
astroid==3.3.10
diff-match-patch==20241021
dill==0.4.0
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
django-import-export==4.3.7
django-js-asset==3.1.2
django-mptt==0.17.0
django-simple-history==3.10.1
djangorestframework==3.16.0
djangorestframework-recursive==0.1.2
et_xmlfile==2.0.0
isort==6.0.1
Markdown==3.8.2
mccabe==0.7.0
numpy==2.3.1
openpyxl==3.1.5
platformdirs==4.3.8
pygraphviz==1.14
pylint==3.3.7
sqlparse==0.5.3
tablib==3.8.0
tomlkit==0.13.3
